refactor: split SQLite backend queries into owned modules
Extract 2431-LOC async_sqlite.py god class into 8 focused query modules
under storage/backends/queries/:
  conversations.py (555), messages.py (265), attachments.py (303),
  raw.py (439), stats.py (256), filter_builder.py (180),
  mappers.py (150), runs.py (61)
async_sqlite.py is now a thin coordinator (~978 LOC).

Also:
- Add storage/hydrators.py: row-to-domain-object converters, breaking
  the circular lib/ ↔ storage/ dependency (L2 domain model inversion)
- Add storage/repair.py: extract repair ops from health.py (798→428 LOC)
- Fix fts5.py: CREATE TABLE used 'content' column, INSERTs used 'text' —
  caused silent data loss on index rebuilds (S6)
- Fix hybrid search: limit=0 now returns [], RRF deduplicates within list
